cloud-run/requirements.txt
firebase-admin
python-docx
google-cloud-storage
google-cloud-firestore
Flask
gunicorn
google-events